improve article-structure plugin types
