+++ addon/components/article-structure-plugin/article-structure-card.ts
@@ -3,19 +3,16 @@ import { tracked } from '@glimmer/tracking';
 import { action } from '@ember/object';
 import { inject as service } from '@ember/service';
 import { ProseController } from '@lblod/ember-rdfa-editor';
-import { Structure } from '@lblod/ember-rdfa-editor-lblod-plugins/utils/article-structure-plugin/constants';
 import { insertArticleStructure } from '@lblod/ember-rdfa-editor-lblod-plugins/commands/article-structure-plugin';
 import IntlService from 'ember-intl/services/intl';
 import { trackedFunction } from 'ember-resources/util/function';
 import validateDatastore from '@lblod/ember-rdfa-editor-lblod-plugins/utils/article-structure-plugin/validate-datastore';
+import { ResolvedArticleStructurePluginOptions } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/article-structure-plugin';
 
 type Args = {
   controller: ProseController;
   widgetArgs: {
-    options: {
-      structures: Structure[];
-      structureTypes: string[];
-    };
+    options: ResolvedArticleStructurePluginOptions;
   };
 };
 export default class EditorPluginsArticleStructureCardComponent extends Component<Args> {
